requirements.txt
requests
chardet
pdfplumber
PyMuPDF
pytesseract
pillow
python-pptx
beautifulsoup4
python-docx
docx2txt
